requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
idna==2.10
mysqlclient==2.0.3
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
